added file log for sender, start rework tests
