+++ hw12_13_14_15_calendar/configs/docker/sender.toml
@@ -9,3 +9,6 @@ topic = "notifications"
 version = "3.2.0"
 maxConnectionTries = 5
 connectionTryDelay = "3s"
+
+[file]
+path =  "./logs/email.log"
